Allow drag and drop of a profile into already-loaded profiler
