+++ src/components/app/Home.css
@@ -27,6 +27,7 @@
   opacity: 0;
   pointer-events: none;
   transition: opacity 200ms;
+  z-index: 10;
 }
 
 .homeDropMessage {
