+++ src/components/app/ProfileViewer.js
@@ -12,6 +12,7 @@ import MenuButtons from './MenuButtons';
 import WindowTitle from '../shared/WindowTitle';
 import SymbolicationStatusOverlay from './SymbolicationStatusOverlay';
 import { returnToZipFileList } from '../../actions/zipped-profiles';
+import { retrieveProfileFromFile } from '../../actions/receive-profile';
 import { getProfileName } from '../../selectors/url-state';
 import Timeline from '../timeline';
 import { getHasZipFile } from '../../selectors/zipped-profiles';
@@ -25,6 +26,7 @@ import {
   getHasSanitizedProfile,
 } from '../../selectors/publish';
 import classNames from 'classnames';
+import DragAndDrop from './DragAndDrop';
 
 import type { CssPixels } from '../../types/units';
 import type { ConnectedProps } from '../../utils/connect';
@@ -44,6 +46,7 @@ type StateProps = {|
 type DispatchProps = {|
   +returnToZipFileList: typeof returnToZipFileList,
   +invalidatePanelLayout: typeof invalidatePanelLayout,
+  +retrieveProfileFromFile: typeof retrieveProfileFromFile,
 |};
 
 type Props = ConnectedProps<{||}, StateProps, DispatchProps>;
@@ -60,14 +63,16 @@ class ProfileViewer extends PureComponent<Props> {
       uploadProgress,
       isHidingStaleProfile,
       hasSanitizedProfile,
+      retrieveProfileFromFile,
     } = this.props;
 
     return (
-      <div
+      <DragAndDrop
         className={classNames({
           profileViewerWrapper: true,
           profileViewerWrapperBackground: hasSanitizedProfile,
         })}
+        retrieveProfileFromFile={retrieveProfileFromFile}
       >
         <div
           className={classNames({
@@ -127,7 +132,7 @@ class ProfileViewer extends PureComponent<Props> {
           <WindowTitle />
           <SymbolicationStatusOverlay />
         </div>
-      </div>
+      </DragAndDrop>
     );
   }
 }
@@ -145,6 +150,7 @@ export default explicitConnect<{||}, StateProps, DispatchProps>({
   mapDispatchToProps: {
     returnToZipFileList,
     invalidatePanelLayout,
+    retrieveProfileFromFile,
   },
   component: ProfileViewer,
 });
